// src/util/types.ts
export enum Wallet {
  MetaMask = 'MetaMask',
  WalletConnect = 'WalletConnect',
  CreateNewWallet = 'Use Email Address to Create Your Wallet',
}